requirements.txt
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.4.0
pandas==2.3.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
rich==14.2.0
six==1.17.0
tzdata==2025.3
pyyaml
pytest